Refactor sampling
