+++ src/mnist/sample.py
@@ -3,74 +3,28 @@ import numpy as np
 
 import torch
 import torch.nn.functional as F
-from torchvision import transforms
 from torchvision.utils import save_image
 
 from data import MNISTDataLoader
-from utils import load_checkpoint
-
-
-
-def fetch_image(label):
-
-    transform = transforms.Compose([transforms.ToTensor()])
-    mnist_data_loader = MNISTDataLoader(transform=transform, shuffle=True)
-
-    mnist_dataset = mnist_data_loader.get_test_data_loader()
-
-    images = mnist_dataset.test_data.numpy()
-    labels = mnist_dataset.test_labels.numpy()
-    images = images[labels == label]
-    image  = images[np.random.choice(np.arange(images.shape[0]))]
-    image  = torch.from_numpy(image).float()
-    image  = image.unsqueeze(0)
-    return image
-
-def fetch_label(label):
-
-    label = torch.LongTensor([label])
-    return label
+from utils import load_checkpoint, check_modality_cond
 
 
 
 if __name__=="__main__":
+  if os.path.basename(os.getcwd()) == 'MVAE':
+    model_path = './src/mnist/trained_models/final_best_epoch.pth.tar'
+  
+  elif os.path.basename(os.getcwd()) == 'src':
+    model_path = './trained_models/final_best_epoch.pth.tar'
 
   condition_on_image = None
-  condition_on_text = 5
+  condition_on_text = 6
   n_samples = 64
-
-  model_path = './trained_models/final_best_epoch.pth.tar'
+  
   model, _, checkpoint = load_checkpoint(model_path)
   model.eval()
 
-  if not condition_on_image and not condition_on_text:
-    mu = torch.Tensor([0])
-    std = torch.Tensor([1])
-
-  elif condition_on_image and not condition_on_text:
-    image = fetch_image(condition_on_image)
-    tmp_mu, tmp_logvar = model.prepare_poe(image_modal=image, label_modal=None)
-
-    mu, logvar = model.compute_poe(tmp_mu, tmp_logvar)
-    std = logvar.mul(0.5).exp_()
-
-  elif condition_on_text and not condition_on_image:
-    label = fetch_label(condition_on_text)
-
-    tmp_mu, tmp_logvar = model.prepare_poe(label_modal=label, image_modal=None)
-
-    mu, logvar = model.compute_poe(tmp_mu, tmp_logvar)
-
-    std = logvar.mul(0.5).exp_()
-
-  elif condition_on_text and condition_on_image:
-      image = fetch_image(condition_on_image)
-      label = fetch_label(condition_on_text)
-
-      tmp_mu, tmp_logvar = model.prepare_poe(image_modal=image, label_modal=label)
-
-      mu, logvar = model.compute_poe(tmp_mu, tmp_logvar)
-      std = logvar.mul(0.5).exp_()
+  mu, std, _, _ = check_modality_cond(condition_on_image, condition_on_text, model)
 
   samples = torch.randn(n_samples, checkpoint["n_latents"])
 
@@ -82,9 +36,14 @@ if __name__=="__main__":
   gen_label  = F.log_softmax(model.label_decoder(samples), dim=1).data
 
 
+  image_data = './sampeled_data'
+  if not os.path.exists(image_data):
+      os.makedirs(image_data)
+  image_path = os.path.join(image_data, 'generated_image.png')
+  
+  save_image(gen_image.view(n_samples, 1, 28, 28), image_path)
 
-  save_image(gen_image.view(n_samples, 1, 28, 28),'./imgs/generated_image.png')
-  with open('./imgs/generated_label.txt', 'w') as fp:
+  with open(os.path.join(image_data, 'generated_label.txt'), 'w') as fp:
       label_numpy = gen_label.numpy()
       label_numpy = np.argmax(label_numpy, axis=1).tolist()
       for i, item in enumerate(label_numpy):
